page headings and objects not traits for components
