+++ app/views/uniform/checkboxes.scala.html
@@ -20,9 +20,10 @@
 @import sdil.config.AppConfig
 @import views.html.helper.options
 @import views.html.uniform.helpers._
+@import ltbs.play._
 
 @(key: String, form: Form[List[String]], possibleValues: List[String], breadcrumbs: List[String])(implicit messages: Messages, request: Request[_], config: AppConfig)
-@views.html.main_template(Messages("sdil.variations.title"), form = Some(form)) {
+@views.html.main_template(messages.first(s"heading.$key", "sdil.variations.title")(), form = Some(form)) {
 
     @backlink(breadcrumbs)
