+++ app/views/uniform/date.scala.html
@@ -21,10 +21,11 @@
 @import sdil.config.AppConfig
 @import views.html.uniform.helpers._
 @import java.time.LocalDate
+@import ltbs.play._
 
 @(key: String, form: Form[LocalDate], breadcrumbs: List[String])(implicit messages: Messages, request: Request[_], config: AppConfig)
 
-@views.html.main_template(Messages("sdil.variations.title"), form = Some(form)) {
+@views.html.main_template(messages.first(s"heading.$key", "sdil.variations.title")(), form = Some(form)) {
 
     @backlink(breadcrumbs)
