+++ app/views/uniform/tellTable.scala.html
@@ -21,9 +21,10 @@
 @import views.html.helper.options
 @import ltbs.play.scaffold.webmonad._
 @import views.html.uniform.helpers._
+@import ltbs.play._
 
 @(key: String, form: Form[Unit], breadcrumbs: List[String], headings: List[Html], rows: List[List[Html]])(implicit messages: Messages, request: Request[_], config: AppConfig)
-@views.html.main_template(Messages("sdil.variations.title"), form = Some(form)) {
+@views.html.main_template(messages.first(s"heading.$key", "sdil.variations.title")(), form = Some(form)) {
 
     @backlink(breadcrumbs)
